Renamed the Editor and Renderer components to PageEditor/-Renderer to fully convey their scope
